Add support for nested wildcards in nested arrays.
Extracted the error sorting test suites of valitate_test into it's own validate_errorSorting_test.js suite to keep things tidier.
